+++ vite.config.js
@@ -1,14 +1,34 @@
 import { defineConfig } from "vite";
 import vue from "@vitejs/plugin-vue";
 import { fileURLToPath, URL } from "node:url";
-// import monacoEditorEsmPlugin from "vite-plugin-monaco-editor-esm";
+import { features } from "monaco-editor/esm/metadata.js";
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [
     vue(),
-    // monacoEditorEsmPlugin({
-    //   languageWorkers: ["editorWorkerService", "typescript"],
-    // }),
+    {
+      name: "use-features",
+      resolveId(id) {
+        if (id === "use-features") {
+          return id; // 标记这是一个虚拟模块
+        }
+      },
+      load(id) {
+        if (id === "use-features") {
+          const prefix = "monaco-editor/esm/";
+          const autoImport = features.map((feat) => {
+            let entry =
+              Object.prototype.toString.call(feat.entry) === "[object String]"
+                ? [feat.entry]
+                : feat.entry;
+            entry = entry.map((val) => `import '${prefix}${val}'`);
+            return `// ${feat.label}
+          ${entry.join("\n")}`;
+          });
+          return autoImport.join("\n");
+        }
+      },
+    },
   ],
   resolve: {
     alias: {