+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### v1.0.4 (_2025-09-30_)
+
+- feat: monaco-editor 打包优化，原来 18MB 减少到现在 11MB，减少了约 38.89%
+- chore(deps): 删除 vite-plugin-monaco-editor-esm 依赖
+- chore(deps): 删除 esbuild 依赖
+
 ### v1.0.3 (_2025-09-30_)
 
 - feat: 编辑 prettier.config.js 可修改 Prettier 配置文件